Renamed userValidations to validations. Update imports
